+++ src/App.vue
@@ -1,5 +1,12 @@
 <template>
-
+  <tyh-card>
+    <template v-slot:title>主标题</template>
+    <template v-slot:subtitle>副标题</template>
+    <p>这是内容部分1</p>
+    <p>这是内容部分2</p>
+    <p>这是内容部分3</p>
+    <p>这是内容部分4</p>
+  </tyh-card>
 </template>
 
 <script setup>