+++ src/main.js
@@ -6,8 +6,8 @@ import router from './router'
 // import tyhUi2 from 'tyh-ui2' 
 // import 'tyh-ui2/style/index.css'
 
-import tyhUi2 from './components/tyhUi'
-import './components/tyhUi/style/index.css'
+import tyhUi2 from './package/tyhUi'
+import './package/tyhUi/style/index.css'
 
 import 'element3/lib/theme-chalk/index.css'
 import Element3 from 'element3'