+++ src/package/tyhUi/packages/card/TyhCard.vue
@@ -1,17 +1,14 @@
 <template>
   <div :class="['tyh-card', { [`tyh-card-shadow-${shadow}`]: shadow }]">
-    <div v-if="!simple" class="tyh-card-header">
-      <p class="tyh-card-title">
+    <div v-if="$slots.title || $slots.subtitle" class="tyh-card-header">
+      <p v-if="$slots.title" class="tyh-card-title">
         <slot name="title" />
       </p>
-      <p class="tyh-card-subtitle">
+      <p v-if="$slots.subtitle" class="tyh-card-subtitle">
         <slot name="subtitle" />
       </p>
     </div>
-
-    <div class="tyh-card-content">
-      <slot />
-    </div>
+    <div class="tyh-card-body" :style="bodyStyle"><slot /></div>
   </div>
 </template>
 
@@ -23,6 +20,6 @@ defineProps({
       return ['always', 'hover', 'noShadow'].includes(v)
     }
   },
-  simple: Boolean
+  bodyStyle: Object
 })
 </script>