+++ src/package/tyhUi/src/index.js
@@ -86,7 +86,7 @@ const install = app => {
 }
 
 const tyhUi = {
-  version: "3.2.18",
+  version: "3.2.19",
   install
 }